requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
accelerate>=0.27.0
sentencepiece>=0.1.99
protobuf>=4.25.0
sentence-transformers>=2.2.0
xgboost>=1.7.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0
huggingface-hub>=0.20.0
